6 - SQM based on JPA type system
