+++ hibernate-core/src/test/java/org/hibernate/test/readonly/ReadOnlySessionLazyNonLazyTest.java
@@ -10,16 +10,15 @@ import java.math.BigDecimal;
 import java.util.Arrays;
 import java.util.HashSet;
 import java.util.Iterator;
-import java.util.List;
 import java.util.Set;
-
-import org.junit.Test;
 
 import org.hibernate.Hibernate;
 import org.hibernate.Session;
 import org.hibernate.Transaction;
 import org.hibernate.engine.spi.SessionImplementor;
 import org.hibernate.proxy.HibernateProxy;
+
+import org.junit.Test;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
@@ -972,340 +971,6 @@ public class ReadOnlySessionLazyNonLazyTest extends AbstractReadOnlyTest {
 		DataPoint dp = ( DataPoint ) s.createQuery( "select c.lazyDataPoints from Container c join c.lazyDataPoints where c.id=" + cOrig.getId() )
 				.setReadOnly( true ).uniqueResult();
 		assertTrue( s.isReadOnly( dp ) );
-		t.commit();
-		s.close();
-		s = openSession();
-		t = s.beginTransaction();
-		s.createQuery("delete from DataPoint").executeUpdate();
-		s.createQuery("delete from Container").executeUpdate();
-		s.createQuery("delete from Info").executeUpdate();
-		s.createQuery("delete from Owner").executeUpdate();
-		t.commit();
-		s.close();
-	}
-
-	@Test
-	@SuppressWarnings( {"unchecked"})
-	public void testDefaultReadOnlyWithModifiableFilterCollectionEntities() {
-		Container cOrig = createContainer();
-		Set expectedInitializedObjects = new HashSet(
-				Arrays.asList(
-						cOrig,
-						cOrig.getNoProxyInfo(),
-						cOrig.getProxyInfo(),
-						cOrig.getNonLazyInfo(),
-						cOrig.getNoProxyOwner(),
-						cOrig.getProxyOwner(),
-						cOrig.getNonLazyOwner(),
-						cOrig.getLazyDataPoints().iterator().next(),
-						cOrig.getNonLazyJoinDataPoints().iterator().next(),
-						cOrig.getNonLazySelectDataPoints().iterator().next()
-				)
-		);
-		Set expectedReadOnlyObjects = new HashSet();
-
-		Session s = openSession();
-		assertFalse( s.isDefaultReadOnly() );
-		Transaction t = s.beginTransaction();
-		s.save( cOrig );
-		checkContainer( cOrig, expectedInitializedObjects, expectedReadOnlyObjects, s );
-		s.setDefaultReadOnly( true );
-		assertTrue( s.isDefaultReadOnly() );
-		checkContainer( cOrig, expectedInitializedObjects, expectedReadOnlyObjects, s );
-		t.commit();
-		s.close();
-
-		s = openSession();
-		t = s.beginTransaction();
-		s.setDefaultReadOnly( true );
-		assertTrue( s.isDefaultReadOnly() );
-		Container c = ( Container ) s.get( Container.class, cOrig.getId() );
-		assertNotSame( cOrig, c );
-		expectedInitializedObjects = new HashSet(
-				Arrays.asList(
-						c,
-						c.getNonLazyInfo(),
-						c.getNoProxyOwner(),
-						c.getProxyOwner(),
-						c.getNonLazyOwner(),
-						c.getNonLazyJoinDataPoints().iterator().next(),
-						c.getNonLazySelectDataPoints().iterator().next()
-				)
-		);
-		expectedReadOnlyObjects = new HashSet(
-				Arrays.asList(
-						c,
-						c.getNoProxyInfo(),
-						c.getProxyInfo(),
-						c.getNonLazyInfo(),
-						c.getNoProxyOwner(),
-						c.getProxyOwner(),
-						c.getNonLazyOwner(),
-						//c.getLazyDataPoints(),
-						c.getNonLazyJoinDataPoints().iterator().next(),
-						c.getNonLazySelectDataPoints().iterator().next()
-				)
-		);
-		List list = s.createFilter( c.getLazyDataPoints(), "" )
-				.setMaxResults(1)
-				.setReadOnly( false )
-				.list();
-		assertEquals( 1, list.size() );
-		assertFalse( s.isReadOnly( list.get( 0 ) ) );
-		list = s.createFilter( c.getNonLazyJoinDataPoints(), "" )
-				.setMaxResults(1)
-				.setReadOnly( false )
-				.list();
-		assertEquals( 1, list.size() );
-		assertTrue( s.isReadOnly( list.get( 0 ) ) );
-		list = s.createFilter( c.getNonLazySelectDataPoints(), "" )
-				.setMaxResults(1)
-				.setReadOnly( false )
-				.list();
-		assertEquals( 1, list.size() );
-		assertTrue( s.isReadOnly( list.get( 0 ) ) );
-		t.commit();
-		s.close();
-		s = openSession();
-		t = s.beginTransaction();
-		s.createQuery("delete from DataPoint").executeUpdate();
-		s.createQuery("delete from Container").executeUpdate();
-		s.createQuery("delete from Info").executeUpdate();
-		s.createQuery("delete from Owner").executeUpdate();
-		t.commit();
-		s.close();
-	}
-
-	@SuppressWarnings( {"unchecked"})
-	@Test
-	public void testDefaultModifiableWithReadOnlyFilterCollectionEntities() {
-		Container cOrig = createContainer();
-		Set expectedInitializedObjects = new HashSet(
-				Arrays.asList(
-						cOrig,
-						cOrig.getNoProxyInfo(),
-						cOrig.getProxyInfo(),
-						cOrig.getNonLazyInfo(),
-						cOrig.getNoProxyOwner(),
-						cOrig.getProxyOwner(),
-						cOrig.getNonLazyOwner(),
-						cOrig.getLazyDataPoints().iterator().next(),
-						cOrig.getNonLazyJoinDataPoints().iterator().next(),
-						cOrig.getNonLazySelectDataPoints().iterator().next()
-				)
-		);
-		Set expectedReadOnlyObjects = new HashSet();
-
-		Session s = openSession();
-		assertFalse( s.isDefaultReadOnly() );
-		Transaction t = s.beginTransaction();
-		s.save( cOrig );
-		checkContainer( cOrig, expectedInitializedObjects, expectedReadOnlyObjects, s );
-		s.setDefaultReadOnly( true );
-		assertTrue( s.isDefaultReadOnly() );
-		checkContainer( cOrig, expectedInitializedObjects, expectedReadOnlyObjects, s );
-		t.commit();
-		s.close();
-
-		s = openSession();
-		t = s.beginTransaction();
-		assertFalse( s.isDefaultReadOnly() );
-		Container c = ( Container ) s.get( Container.class, cOrig.getId() );
-		assertNotSame( cOrig, c );
-		expectedInitializedObjects = new HashSet(
-				Arrays.asList(
-						c,
-						c.getNonLazyInfo(),
-						c.getNoProxyOwner(),
-						c.getProxyOwner(),
-						c.getNonLazyOwner(),
-						c.getNonLazyJoinDataPoints().iterator().next(),
-						c.getNonLazySelectDataPoints().iterator().next()
-				)
-		);
-		expectedReadOnlyObjects = new HashSet();
-		List list = s.createFilter( c.getLazyDataPoints(), "" )
-				.setMaxResults(1)
-				.setReadOnly( true )
-				.list();
-		assertEquals( 1, list.size() );
-		assertTrue( s.isReadOnly( list.get( 0 ) ) );
-		list = s.createFilter( c.getNonLazyJoinDataPoints(), "" )
-				.setMaxResults(1)
-				.setReadOnly( true )
-				.list();
-		assertEquals( 1, list.size() );
-		assertFalse( s.isReadOnly( list.get( 0 ) ) );
-		list = s.createFilter( c.getNonLazySelectDataPoints(), "" )
-				.setMaxResults(1)
-				.setReadOnly( true )
-				.list();
-		assertEquals( 1, list.size() );
-		assertFalse( s.isReadOnly( list.get( 0 ) ) );
-		t.commit();
-		s.close();
-		s = openSession();
-		t = s.beginTransaction();
-		s.createQuery("delete from DataPoint").executeUpdate();
-		s.createQuery("delete from Container").executeUpdate();
-		s.createQuery("delete from Info").executeUpdate();
-		s.createQuery("delete from Owner").executeUpdate();
-		t.commit();
-		s.close();
-	}
-
-	@SuppressWarnings( {"unchecked"})
-	@Test
-	public void testDefaultReadOnlyWithFilterCollectionEntities() {
-		Container cOrig = createContainer();
-		Set expectedInitializedObjects = new HashSet(
-				Arrays.asList(
-						cOrig,
-						cOrig.getNoProxyInfo(),
-						cOrig.getProxyInfo(),
-						cOrig.getNonLazyInfo(),
-						cOrig.getNoProxyOwner(),
-						cOrig.getProxyOwner(),
-						cOrig.getNonLazyOwner(),
-						cOrig.getLazyDataPoints().iterator().next(),
-						cOrig.getNonLazyJoinDataPoints().iterator().next(),
-						cOrig.getNonLazySelectDataPoints().iterator().next()
-				)
-		);
-		Set expectedReadOnlyObjects = new HashSet();
-
-		Session s = openSession();
-		assertFalse( s.isDefaultReadOnly() );
-		Transaction t = s.beginTransaction();
-		s.save( cOrig );
-		checkContainer( cOrig, expectedInitializedObjects, expectedReadOnlyObjects, s );
-		s.setDefaultReadOnly( true );
-		assertTrue( s.isDefaultReadOnly() );
-		checkContainer( cOrig, expectedInitializedObjects, expectedReadOnlyObjects, s );
-		t.commit();
-		s.close();
-
-		s = openSession();
-		t = s.beginTransaction();
-		s.setDefaultReadOnly( true );
-		assertTrue( s.isDefaultReadOnly() );
-		Container c = ( Container ) s.get( Container.class, cOrig.getId() );
-		assertNotSame( cOrig, c );
-		expectedInitializedObjects = new HashSet(
-				Arrays.asList(
-						c,
-						c.getNonLazyInfo(),
-						c.getNoProxyOwner(),
-						c.getProxyOwner(),
-						c.getNonLazyOwner(),
-						c.getNonLazyJoinDataPoints().iterator().next(),
-						c.getNonLazySelectDataPoints().iterator().next()
-				)
-		);
-		expectedReadOnlyObjects = new HashSet(
-				Arrays.asList(
-						c,
-						c.getNoProxyInfo(),
-						c.getProxyInfo(),
-						c.getNonLazyInfo(),
-						c.getNoProxyOwner(),
-						c.getProxyOwner(),
-						c.getNonLazyOwner(),
-						//c.getLazyDataPoints(),
-						c.getNonLazyJoinDataPoints().iterator().next(),
-						c.getNonLazySelectDataPoints().iterator().next()
-				)
-		);
-		List list = s.createFilter( c.getLazyDataPoints(), "" )
-				.setMaxResults( 1 )
-				.list();
-		assertEquals( 1, list.size() );
-		assertTrue( s.isReadOnly( list.get( 0 ) ) );
-		list = s.createFilter( c.getNonLazyJoinDataPoints(), "" )
-				.setMaxResults( 1 )
-				.list();
-		assertEquals( 1, list.size() );
-		assertTrue( s.isReadOnly( list.get( 0 ) ) );
-		list = s.createFilter( c.getNonLazySelectDataPoints(), "" )
-				.setMaxResults( 1 )
-				.list();
-		assertEquals( 1, list.size() );
-		assertTrue( s.isReadOnly( list.get( 0 ) ) );
-		t.commit();
-		s.close();
-		s = openSession();
-		t = s.beginTransaction();
-		s.createQuery("delete from DataPoint").executeUpdate();
-		s.createQuery("delete from Container").executeUpdate();
-		s.createQuery("delete from Info").executeUpdate();
-		s.createQuery("delete from Owner").executeUpdate();
-		t.commit();
-		s.close();
-	}
-
-	@SuppressWarnings( {"unchecked"})
-	@Test
-	public void testDefaultModifiableWithFilterCollectionEntities() {
-		Container cOrig = createContainer();
-		Set expectedInitializedObjects = new HashSet(
-				Arrays.asList(
-						cOrig,
-						cOrig.getNoProxyInfo(),
-						cOrig.getProxyInfo(),
-						cOrig.getNonLazyInfo(),
-						cOrig.getNoProxyOwner(),
-						cOrig.getProxyOwner(),
-						cOrig.getNonLazyOwner(),
-						cOrig.getLazyDataPoints().iterator().next(),
-						cOrig.getNonLazyJoinDataPoints().iterator().next(),
-						cOrig.getNonLazySelectDataPoints().iterator().next()
-				)
-		);
-		Set expectedReadOnlyObjects = new HashSet();
-
-		Session s = openSession();
-		assertFalse( s.isDefaultReadOnly() );
-		Transaction t = s.beginTransaction();
-		s.save( cOrig );
-		checkContainer( cOrig, expectedInitializedObjects, expectedReadOnlyObjects, s );
-		s.setDefaultReadOnly( true );
-		assertTrue( s.isDefaultReadOnly() );
-		checkContainer( cOrig, expectedInitializedObjects, expectedReadOnlyObjects, s );
-		t.commit();
-		s.close();
-
-		s = openSession();
-		t = s.beginTransaction();
-		assertFalse( s.isDefaultReadOnly() );
-		Container c = ( Container ) s.get( Container.class, cOrig.getId() );
-		assertNotSame( cOrig, c );
-		expectedInitializedObjects = new HashSet(
-				Arrays.asList(
-						c,
-						c.getNonLazyInfo(),
-						c.getNoProxyOwner(),
-						c.getProxyOwner(),
-						c.getNonLazyOwner(),
-						c.getNonLazyJoinDataPoints().iterator().next(),
-						c.getNonLazySelectDataPoints().iterator().next()
-				)
-		);
-		expectedReadOnlyObjects = new HashSet();
-		List list = s.createFilter( c.getLazyDataPoints(), "" )
-				.setMaxResults( 1 )
-				.list();
-		assertEquals( 1, list.size() );
-		assertFalse( s.isReadOnly( list.get( 0 ) ) );
-		list = s.createFilter( c.getNonLazyJoinDataPoints(), "" )
-				.setMaxResults( 1 )
-				.list();
-		assertEquals( 1, list.size() );
-		assertFalse( s.isReadOnly( list.get( 0 ) ) );
-		list = s.createFilter( c.getNonLazySelectDataPoints(), "" )
-				.setMaxResults( 1 )
-				.list();
-		assertEquals( 1, list.size() );
-		assertFalse( s.isReadOnly( list.get( 0 ) ) );
 		t.commit();
 		s.close();
 		s = openSession();
